Fix translucent navigation bar layout on tablets in landscape.
The status bar blocker view was visible and stretched over the whole
navigation bar, even though it's not even needed on tablets, as the
status bar never extends into the navigation bar there. Fix that by
hiding the blocker view, but keep it in the layout in oder to not
clutter the code with null checks.

Also cleanup unneeded method override.

Change-Id: I50d58e6e3496516b37f8c36b497cf8c9f569325c

Conflicts:
	packages/SystemUI/res/layout-sw600dp/navigation_bar.xml